Refactor: Post 제목과 내용 검증 로직을 service에서 domain으로 이동
